+++ controllers/news_v2.go
@@ -2,17 +2,18 @@ package controllers
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 	"sort"
+	"strconv"
 	"sync"
-	"fmt"
 
 	"github.com/gin-gonic/gin"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
-	f "github.com/twreporter/logformatter"
 	"github.com/twreporter/go-api/globals"
 	"github.com/twreporter/go-api/internal/news"
+	f "github.com/twreporter/logformatter"
 )
 
 type newsV2Storage interface {
@@ -65,14 +66,16 @@ func (nc *newsV2Controller) GetPosts(c *gin.Context) {
 		return
 	}
 
+	toggleBookmark, _ := strconv.Atoi(c.Query("toggleBookmark"))
+	if toggleBookmark == 1 {
+		c.Writer.Header().Set("Cache-Control", "no-store")
+	}
 	authUserID := c.Request.Context().Value(globals.AuthUserIDProperty)
-	if authUserID != nil {
+	if authUserID != nil && toggleBookmark == 1 {
 		authUserIdString := fmt.Sprintf("%v", authUserID)
 		if _, err := nc.SqlStorage.GetBookmarksOfPosts(ctx, authUserIdString, posts); err != nil {
 			log.WithField("detail", err).Errorf("%s", f.FormatStack(err))
 		}
-
-		c.Writer.Header().Set("Cache-Control", "no-cache")
 	}
 
 	total, err := nc.Storage.GetPostCount(ctx, q)
@@ -319,13 +322,13 @@ func (nc *newsV2Controller) getIndexPageJobs() []job {
 			Name: news.ReviewsSection,
 			Type: typePost,
 			Query: news.NewQuery(
-				news.WithFilterCategoryIDs(news.Opinion.Key),
+				news.WithFilterCategorySet(news.Opinion.Key),
 				news.WithLimit(4)),
 		}, {
 			Name: news.PhotoSection,
 			Type: typePost,
 			Query: news.NewQuery(
-				news.WithFilterCategoryIDs(news.Photography.ID),
+				news.WithFilterCategorySet(news.Photography.ID),
 				news.WithLimit(6)),
 		}, {
 			Name:  news.InfographicSection,
@@ -336,22 +339,6 @@ func (nc *newsV2Controller) getIndexPageJobs() []job {
 			Type:  typeTopic,
 			Query: news.NewQuery(news.WithOffset(1), news.WithLimit(4)),
 		},
-	}
-
-	// v1 categories in index page
-	for _, v := range []news.Category{
-		news.HumanRightsAndSociety,
-		news.EnvironmentAndEducation,
-		news.PoliticsAndEconomy,
-		news.CultureAndArt,
-		news.International,
-		news.LivingAndMedicalCare,
-	} {
-		jobs = append(jobs, job{
-			v.Name,
-			typePost,
-			news.NewQuery(news.WithFilterCategoryIDs(v.ID), news.WithLimit(1)),
-		})
 	}
 
 	// v2 categories in index page